save enemy

+++ core/src/com/group12/cookiesrising/Player.java
@@ -189,7 +189,7 @@ public class Player extends Observable implements Upgradable,Observer,Health,Hit
     }
 
     public void loadData(){
-        if(SaveManager.loadDataValue("money",int.class) != null){
+        if(SaveManager.loadDataValue("money",double.class) != null){
             money = SaveManager.loadDataValue("money",int.class);
         } else{
             money = 0;
